pyproject.toml
[tool.poetry]
name = "bank-ledger-9000"
version = "0.1.0"
description = "Group project to make bank system"
authors = ["Your Name <you@example.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-two-factor-auth = "^1.14.0"
phonenumbers = "^8.13.0"
django-rq = "^2.6.0"
requests = "^2.28.1"
django-crontab = "^0.7.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
